rspirv::binary

Module disassemble

Macros

push 🔒

Pushes the given value to the given container if the value is not empty.

Traits

Disassemble

Trait for disassembling functionalities.

DisassembleLiteralBit 🔒

Functions

disas_constant 🔒

disas_ext_inst 🔒

disas_instruction 🔒

disas_join 🔒

Disassembles each instruction in insts and joins them together with the given delimiter.

disas_literal_bit_operand 🔒