Enum rspirv::binary::ParseState

pub enum ParseState {

    Complete,
    ConsumerStopRequested,
    ConsumerError(Box<dyn Error + Send>),
    HeaderIncomplete(DecodeError),
    HeaderIncorrect,
    EndiannessUnsupported,
    WordCountZero(usize, usize),
    OpcodeUnknown(usize, usize, u16),
    OperandExpected(usize, usize),
    OperandExceeded(usize, usize),
    OperandError(DecodeError),
    TypeUnsupported(usize, usize),
    SpecConstantOpIntegerIncorrect(usize, usize),

}

Parser State.

Most of the error variants will retain the error location for both byte offset (starting from 0) and instruction number (starting from 1).

Variants

Complete

Parsing completed

ConsumerStopRequested

Consumer requested to stop parse

ConsumerError(Box<dyn Error + Send>)

Consumer errored out with the given error

HeaderIncomplete(DecodeError)

Incomplete module header

HeaderIncorrect

Incorrect module header

EndiannessUnsupported

Unsupported endianness

WordCountZero(usize, usize)

Zero instruction word count at (byte offset, inst number)

OpcodeUnknown(usize, usize, u16)

Unknown opcode at (byte offset, inst number, opcode)

OperandExpected(usize, usize)

Expected more operands (byte offset, inst number)

OperandExceeded(usize, usize)

found redundant operands (byte offset, inst number)

OperandError(DecodeError)

Errored out when decoding operand with the given error

TypeUnsupported(usize, usize)

Unsupported type (byte offset, inst number)

SpecConstantOpIntegerIncorrect(usize, usize)

Incorrect SpecConstantOp Integer (byte offset, inst number)