rspirv::binary

Module parser

Structs

Parser

The SPIR-V binary parser.

Enums

Action

Orders consumer sent to the parser after each consuming call.

State

Parser State.

Constants

HEADER_NUM_WORDS 🔒

WORD_NUM_BYTES 🔒

Traits

Consumer

The binary consumer trait.

Functions

parse_bytes

Parses the given binary and consumes the module using the given consumer.

parse_words

Parses the given binary and consumes the module using the given consumer.

Type Aliases

GInstRef 🔒

Result