rspirv::binary::parser

Struct Parser

pub struct Parser<'c, 'd> {
    decoder: Decoder<'d>,
    consumer: &'c mut dyn Consumer,
    type_tracker: TypeTracker,
    inst_index: usize,
}

The SPIR-V binary parser.

Takes in a vector of bytes and a consumer, this parser will invoke the consume methods on the consumer for the module header and each instruction parsed.

Different from the Decoder, this parser is high-level; it has knowlege of the SPIR-V grammar. It will parse instructions according to SPIR-V grammar.

Examples

use spirv::{AddressingModel, MemoryModel};
use rspirv::binary::Parser;
use rspirv::dr::{Loader, Operand};

let bin = vec![
    // Magic number.           Version number: 1.0.
    0x03, 0x02, 0x23, 0x07,    0x00, 0x00, 0x01, 0x00,
    // Generator number: 0.    Bound: 0.
    0x00, 0x00, 0x00, 0x00,    0x00, 0x00, 0x00, 0x00,
    // Reserved word: 0.
    0x00, 0x00, 0x00, 0x00,
    // OpMemoryModel.          Logical.
    0x0e, 0x00, 0x03, 0x00,    0x00, 0x00, 0x00, 0x00,
    // GLSL450.
    0x01, 0x00, 0x00, 0x00];
let mut loader = Loader::new();  // You can use your own consumer here.
{
    let p = Parser::new(&bin, &mut loader);
    p.parse().unwrap();
}
let module = loader.module();

assert_eq!((1, 0), module.header.unwrap().version());
let m = module.memory_model.as_ref().unwrap();
assert_eq!(Operand::AddressingModel(AddressingModel::Logical),
           m.operands[0]);
assert_eq!(Operand::MemoryModel(MemoryModel::GLSL450),
           m.operands[1]);

Fields
decoder: Decoder<'d>consumer: &'c mut dyn Consumertype_tracker: TypeTrackerinst_index: usize

The index of the current instructions

Starting from 1, 0 means invalid

Implementations

impl<'c, 'd> Parser<'c, 'd>

pub fn new(binary: &'d [u8], consumer: &'c mut dyn Consumer) -> Self

Creates a new parser to parse the given binary and send the module header and instructions to the given consumer.

pub fn parse(self) -> Result<()>

Does the parsing.

fn split_into_word_count_and_opcode(word: Word) -> (u16, u16)

fn parse_header(&mut self) -> Result<ModuleHeader>

fn parse_inst(&mut self) -> Result<Instruction>

fn parse_literal(&mut self, type_id: Word) -> Result<Operand>

fn parse_spec_constant_op(&mut self) -> Result<Vec<Operand>>

fn parse_operands( &mut self, grammar: &'static Instruction<'static>, ) -> Result<Instruction>

impl Parser<'_, '_>

fn parse_operand(&mut self, kind: GOpKind) -> Result<Vec<Operand>>

fn parse_image_operands_arguments( &mut self, image_operands: ImageOperands, ) -> Result<Vec<Operand>>

fn parse_loop_control_arguments( &mut self, loop_control: LoopControl, ) -> Result<Vec<Operand>>

fn parse_memory_access_arguments( &mut self, memory_access: MemoryAccess, ) -> Result<Vec<Operand>>

fn parse_execution_mode_arguments( &mut self, execution_mode: ExecutionMode, ) -> Result<Vec<Operand>>

fn parse_decoration_arguments( &mut self, decoration: Decoration, ) -> Result<Vec<Operand>>

fn parse_tensor_addressing_operands_arguments( &mut self, tensor_addressing_operands: TensorAddressingOperands, ) -> Result<Vec<Operand>>

fn parse_tensor_operands_arguments( &mut self, tensor_operands: TensorOperands, ) -> Result<Vec<Operand>>