rspirv::binary::tracker

Struct ExtInstSetTracker

pub struct ExtInstSetTracker {
    sets: HashMap<Word, &'static InstructionTable<ExtInstOp>>,
}

Struct for tracking extended instruction sets.

If a given extended instruction set is not supported, it will just be silently ignored.

Fields
sets: HashMap<Word, &'static InstructionTable<ExtInstOp>>
Implementations

impl ExtInstSetTracker

pub fn new() -> ExtInstSetTracker

pub fn track(&mut self, inst: &Instruction)

Tracks the extended instruction set declared by the given inst.

If the given extended instruction set is not recognized, it will be silently ignored.

pub fn resolve( &self, set: Word, opcode: Word, ) -> Option<&'static ExtendedInstruction<'static>>

Resolves the extended instruction with opcode in set set.

This method will return None for both untracked instruction sets and unknown opcode in tracked instruction sets.