Struct rspirv::dr::Loader

pub struct Loader { /* private fields */ }

The data representation loader.

Constructs a Module from the module header and instructions.

It implements the Consumer trait and works with the Parser.

Implementations

impl Loader

pub fn new() -> Loader

Creates a new empty loader.

pub fn module(self) -> Module

Returns the Module under construction.