Struct rspirv::dr::Module

pub struct Module {
    pub header: Option<ModuleHeader>,
    pub capabilities: Vec<Instruction>,
    pub extensions: Vec<Instruction>,
    pub ext_inst_imports: Vec<Instruction>,
    pub memory_model: Option<Instruction>,
    pub entry_points: Vec<Instruction>,
    pub execution_modes: Vec<Instruction>,
    pub debugs: Vec<Instruction>,
    pub annotations: Vec<Instruction>,
    pub types_global_values: Vec<Instruction>,
    pub functions: Vec<Function>,
}

Data representation of a SPIR-V module.

Most of the fields are just vectors of Instructions, but some fields store values decomposed from Instructions for better investigation.

The order of its fields basically reveal the requirements in the Logical Layout of a Module of the SPIR-V of the SPIR-V specification.

Fields
header: Option<ModuleHeader>

The module header.

capabilities: Vec<Instruction>

All OpCapability instructions.

extensions: Vec<Instruction>

All OpExtension instructions.

ext_inst_imports: Vec<Instruction>

All OpExtInstImport instructions.

memory_model: Option<Instruction>

The OpMemoryModel instruction.

Although it is required by the specification to appear exactly once per module, we keep it optional here to allow flexibility.

entry_points: Vec<Instruction>

All entry point declarations, using OpEntryPoint.

execution_modes: Vec<Instruction>

All execution mode declarations, using OpExecutionMode.

debugs: Vec<Instruction>

All non-location debug instructions.

annotations: Vec<Instruction>

All annotation instructions.

types_global_values: Vec<Instruction>

All types, constants, and global variables.

As per the specification, they have to be bundled together because they can depend on one another.

functions: Vec<Function>

All functions.

Implementations

impl Module

pub fn new() -> Self

Creates a new empty Module instance.

pub fn global_inst_iter(&self) -> impl Iterator<Item = &Instruction>

Returns an iterator over all global instructions.

pub fn global_inst_iter_mut(&mut self) -> impl Iterator<Item = &mut Instruction>

Returns a mut iterator over all global instructions.

pub fn all_inst_iter(&self) -> impl Iterator<Item = &Instruction>

Returns a iterator over all instructions.

pub fn all_inst_iter_mut(&mut self) -> impl Iterator<Item = &mut Instruction>

Returns a mut iterator over all instructions.