Module rspirv::binary

Module for SPIR-V binary processing.

This module provides a Decoder and a Parser:

The decoder is a low-level binary processing tool; it has no knowlege of the SPIR-V grammar. It only serves SPIR-V word requests.

The parser is a high-level binary processing tool; it has knowledge of the SPIR-V grammar. It works with the Consumer to process a SPIR-V binary on the instruction level.

Structs

Decoder
The SPIR-V binary decoder.

Parser
The SPIR-V binary parser.

Enums

DecodeError
Decoder Error

ParseAction
Orders consumer sent to the parser after each consuming call.

ParseState
Parser State.

Traits

Assemble
Trait for assembling functionalities.

Consumer
The binary consumer trait.

Disassemble
Trait for disassembling functionalities.

Functions

parse_bytes
Parses the given binary and consumes the module using the given consumer.

parse_words
Parses the given binary and consumes the module using the given consumer.

Type Definitions

ParseResult