Enum rspirv::binary::DecodeError

pub enum DecodeError {
    StreamExpected(usize),
    LimitReached(usize),
    ImageOperandsUnknown(usize, Word),
    FPFastMathModeUnknown(usize, Word),
    SelectionControlUnknown(usize, Word),
    LoopControlUnknown(usize, Word),
    FunctionControlUnknown(usize, Word),
    MemorySemanticsUnknown(usize, Word),
    MemoryAccessUnknown(usize, Word),
    KernelProfilingInfoUnknown(usize, Word),
    RayFlagsUnknown(usize, Word),
    SourceLanguageUnknown(usize, Word),
    ExecutionModelUnknown(usize, Word),
    AddressingModelUnknown(usize, Word),
    MemoryModelUnknown(usize, Word),
    ExecutionModeUnknown(usize, Word),
    StorageClassUnknown(usize, Word),
    DimUnknown(usize, Word),
    SamplerAddressingModeUnknown(usize, Word),
    SamplerFilterModeUnknown(usize, Word),
    ImageFormatUnknown(usize, Word),
    ImageChannelOrderUnknown(usize, Word),
    ImageChannelDataTypeUnknown(usize, Word),
    FPRoundingModeUnknown(usize, Word),
    LinkageTypeUnknown(usize, Word),
    AccessQualifierUnknown(usize, Word),
    FunctionParameterAttributeUnknown(usize, Word),
    DecorationUnknown(usize, Word),
    BuiltInUnknown(usize, Word),
    ScopeUnknown(usize, Word),
    GroupOperationUnknown(usize, Word),
    KernelEnqueueFlagsUnknown(usize, Word),
    CapabilityUnknown(usize, Word),
    RayQueryIntersectionUnknown(usize, Word),
    RayQueryCommittedIntersectionTypeUnknown(usize, Word),
    RayQueryCandidateIntersectionTypeUnknown(usize, Word),
    DecodeStringFailed(usize, String),
}

Decoder Error

Variants

StreamExpected(usize)

LimitReached(usize)

ImageOperandsUnknown(usize, Word)

FPFastMathModeUnknown(usize, Word)

SelectionControlUnknown(usize, Word)

LoopControlUnknown(usize, Word)

FunctionControlUnknown(usize, Word)

MemorySemanticsUnknown(usize, Word)

MemoryAccessUnknown(usize, Word)

KernelProfilingInfoUnknown(usize, Word)

RayFlagsUnknown(usize, Word)

SourceLanguageUnknown(usize, Word)

ExecutionModelUnknown(usize, Word)

AddressingModelUnknown(usize, Word)

MemoryModelUnknown(usize, Word)

ExecutionModeUnknown(usize, Word)

StorageClassUnknown(usize, Word)

DimUnknown(usize, Word)

SamplerAddressingModeUnknown(usize, Word)

SamplerFilterModeUnknown(usize, Word)

ImageFormatUnknown(usize, Word)

ImageChannelOrderUnknown(usize, Word)

ImageChannelDataTypeUnknown(usize, Word)

FPRoundingModeUnknown(usize, Word)

LinkageTypeUnknown(usize, Word)

AccessQualifierUnknown(usize, Word)

FunctionParameterAttributeUnknown(usize, Word)

DecorationUnknown(usize, Word)

BuiltInUnknown(usize, Word)

ScopeUnknown(usize, Word)

GroupOperationUnknown(usize, Word)

KernelEnqueueFlagsUnknown(usize, Word)

CapabilityUnknown(usize, Word)

RayQueryIntersectionUnknown(usize, Word)

RayQueryCommittedIntersectionTypeUnknown(usize, Word)

RayQueryCandidateIntersectionTypeUnknown(usize, Word)

DecodeStringFailed(usize, String)

Failed to decode a string.

For structured error handling, the second element could be string::FromUtf8Error, but the will prohibit the compiler from generating PartialEq for this enum.