Enum rspirv::binary::DecodeError

pub enum DecodeError {

    StreamExpected(usize),
    LimitReached(usize),
    ImageOperandsUnknown(usize, Word),
    FPFastMathModeUnknown(usize, Word),
    SelectionControlUnknown(usize, Word),
    LoopControlUnknown(usize, Word),
    FunctionControlUnknown(usize, Word),
    MemorySemanticsUnknown(usize, Word),
    MemoryAccessUnknown(usize, Word),
    KernelProfilingInfoUnknown(usize, Word),
    RayFlagsUnknown(usize, Word),
    FragmentShadingRateUnknown(usize, Word),
    SourceLanguageUnknown(usize, Word),
    ExecutionModelUnknown(usize, Word),
    AddressingModelUnknown(usize, Word),
    MemoryModelUnknown(usize, Word),
    ExecutionModeUnknown(usize, Word),
    StorageClassUnknown(usize, Word),
    DimUnknown(usize, Word),
    SamplerAddressingModeUnknown(usize, Word),
    SamplerFilterModeUnknown(usize, Word),
    ImageFormatUnknown(usize, Word),
    ImageChannelOrderUnknown(usize, Word),
    ImageChannelDataTypeUnknown(usize, Word),
    FPRoundingModeUnknown(usize, Word),
    LinkageTypeUnknown(usize, Word),
    AccessQualifierUnknown(usize, Word),
    FunctionParameterAttributeUnknown(usize, Word),
    DecorationUnknown(usize, Word),
    BuiltInUnknown(usize, Word),
    ScopeUnknown(usize, Word),
    GroupOperationUnknown(usize, Word),
    KernelEnqueueFlagsUnknown(usize, Word),
    CapabilityUnknown(usize, Word),
    RayQueryIntersectionUnknown(usize, Word),
    RayQueryCommittedIntersectionTypeUnknown(usize, Word),
    RayQueryCandidateIntersectionTypeUnknown(usize, Word),
    DecodeStringFailed(usize, String),

}

Decoder Error

Variants

StreamExpected(usize)

Tuple Fields of StreamExpected

0: usize

LimitReached(usize)

Tuple Fields of LimitReached

0: usize

ImageOperandsUnknown(usize, Word)

Tuple Fields of ImageOperandsUnknown

0: usize1: Word

FPFastMathModeUnknown(usize, Word)

Tuple Fields of FPFastMathModeUnknown

0: usize1: Word

SelectionControlUnknown(usize, Word)

Tuple Fields of SelectionControlUnknown

0: usize1: Word

LoopControlUnknown(usize, Word)

Tuple Fields of LoopControlUnknown

0: usize1: Word

FunctionControlUnknown(usize, Word)

Tuple Fields of FunctionControlUnknown

0: usize1: Word

MemorySemanticsUnknown(usize, Word)

Tuple Fields of MemorySemanticsUnknown

0: usize1: Word

MemoryAccessUnknown(usize, Word)

Tuple Fields of MemoryAccessUnknown

0: usize1: Word

KernelProfilingInfoUnknown(usize, Word)

Tuple Fields of KernelProfilingInfoUnknown

0: usize1: Word

RayFlagsUnknown(usize, Word)

Tuple Fields of RayFlagsUnknown

0: usize1: Word

FragmentShadingRateUnknown(usize, Word)

Tuple Fields of FragmentShadingRateUnknown

0: usize1: Word

SourceLanguageUnknown(usize, Word)

Tuple Fields of SourceLanguageUnknown

0: usize1: Word

ExecutionModelUnknown(usize, Word)

Tuple Fields of ExecutionModelUnknown

0: usize1: Word

AddressingModelUnknown(usize, Word)

Tuple Fields of AddressingModelUnknown

0: usize1: Word

MemoryModelUnknown(usize, Word)

Tuple Fields of MemoryModelUnknown

0: usize1: Word

ExecutionModeUnknown(usize, Word)

Tuple Fields of ExecutionModeUnknown

0: usize1: Word

StorageClassUnknown(usize, Word)

Tuple Fields of StorageClassUnknown

0: usize1: Word

DimUnknown(usize, Word)

Tuple Fields of DimUnknown

0: usize1: Word

SamplerAddressingModeUnknown(usize, Word)

Tuple Fields of SamplerAddressingModeUnknown

0: usize1: Word

SamplerFilterModeUnknown(usize, Word)

Tuple Fields of SamplerFilterModeUnknown

0: usize1: Word

ImageFormatUnknown(usize, Word)

Tuple Fields of ImageFormatUnknown

0: usize1: Word

ImageChannelOrderUnknown(usize, Word)

Tuple Fields of ImageChannelOrderUnknown

0: usize1: Word

ImageChannelDataTypeUnknown(usize, Word)

Tuple Fields of ImageChannelDataTypeUnknown

0: usize1: Word

FPRoundingModeUnknown(usize, Word)

Tuple Fields of FPRoundingModeUnknown

0: usize1: Word

LinkageTypeUnknown(usize, Word)

Tuple Fields of LinkageTypeUnknown

0: usize1: Word

AccessQualifierUnknown(usize, Word)

Tuple Fields of AccessQualifierUnknown

0: usize1: Word

FunctionParameterAttributeUnknown(usize, Word)

Tuple Fields of FunctionParameterAttributeUnknown

0: usize1: Word

DecorationUnknown(usize, Word)

Tuple Fields of DecorationUnknown

0: usize1: Word

BuiltInUnknown(usize, Word)

Tuple Fields of BuiltInUnknown

0: usize1: Word

ScopeUnknown(usize, Word)

Tuple Fields of ScopeUnknown

0: usize1: Word

GroupOperationUnknown(usize, Word)

Tuple Fields of GroupOperationUnknown

0: usize1: Word

KernelEnqueueFlagsUnknown(usize, Word)

Tuple Fields of KernelEnqueueFlagsUnknown

0: usize1: Word

CapabilityUnknown(usize, Word)

Tuple Fields of CapabilityUnknown

0: usize1: Word

RayQueryIntersectionUnknown(usize, Word)

Tuple Fields of RayQueryIntersectionUnknown

0: usize1: Word

RayQueryCommittedIntersectionTypeUnknown(usize, Word)

Tuple Fields of RayQueryCommittedIntersectionTypeUnknown

0: usize1: Word

RayQueryCandidateIntersectionTypeUnknown(usize, Word)

Tuple Fields of RayQueryCandidateIntersectionTypeUnknown

0: usize1: Word

DecodeStringFailed(usize, String)

Failed to decode a string.

For structured error handling, the second element could be string::FromUtf8Error, but the will prohibit the compiler from generating PartialEq for this enum.

Tuple Fields of DecodeStringFailed

0: usize1: String