Enum rspirv::binary::ParseState

pub enum ParseState {

    Complete,
    ConsumerStopRequested,
    ConsumerError(Box<dyn Error + Send>),
    HeaderIncomplete(DecodeError),
    HeaderIncorrect,
    EndiannessUnsupported,
    WordCountZero(usize, usize),
    OpcodeUnknown(usize, usize, u16),
    OperandExpected(usize, usize),
    OperandExceeded(usize, usize),
    OperandError(DecodeError),
    TypeUnsupported(usize, usize),
    SpecConstantOpIntegerIncorrect(usize, usize),

}

Parser State.

Most of the error variants will retain the error location for both byte offset (starting from 0) and instruction number (starting from 1).

Variants

Complete

Parsing completed

ConsumerStopRequested

Consumer requested to stop parse

ConsumerError(Box<dyn Error + Send>)

Consumer errored out with the given error

Tuple Fields of ConsumerError

0: Box<dyn Error + Send>

HeaderIncomplete(DecodeError)

Incomplete module header

Tuple Fields of HeaderIncomplete

0: DecodeError

HeaderIncorrect

Incorrect module header

EndiannessUnsupported

Unsupported endianness

WordCountZero(usize, usize)

Zero instruction word count at (byte offset, inst number)

Tuple Fields of WordCountZero

0: usize1: usize

OpcodeUnknown(usize, usize, u16)

Unknown opcode at (byte offset, inst number, opcode)

Tuple Fields of OpcodeUnknown

0: usize1: usize2: u16

OperandExpected(usize, usize)

Expected more operands (byte offset, inst number)

Tuple Fields of OperandExpected

0: usize1: usize

OperandExceeded(usize, usize)

found redundant operands (byte offset, inst number)

Tuple Fields of OperandExceeded

0: usize1: usize

OperandError(DecodeError)

Errored out when decoding operand with the given error

Tuple Fields of OperandError

0: DecodeError

TypeUnsupported(usize, usize)

Unsupported type (byte offset, inst number)

Tuple Fields of TypeUnsupported

0: usize1: usize

SpecConstantOpIntegerIncorrect(usize, usize)

Incorrect SpecConstantOp Integer (byte offset, inst number)

Tuple Fields of SpecConstantOpIntegerIncorrect

0: usize1: usize