Trait rspirv::binary::Assemble

pub trait Assemble {
    pub fn assemble_into(&self, result: &mut Vec<u32>);

    pub fn assemble(&self) -> Vec<u32> { ... }
}

Trait for assembling functionalities.

Required methods

pub fn assemble_into(&self, result: &mut Vec<u32>)

Assembles the current object into the result vector, reducing the need for lots of allocations

Loading content...
Provided methods

pub fn assemble(&self) -> Vec<u32>

Assembles the current object and returns the binary code. Helper method to remain backwards compatible, calls assemble_into

Loading content...