Trait rspirv::binary::Assemble

pub trait Assemble {
    fn assemble_into(&self, result: &mut Vec<u32>);

    fn assemble(&self) -> Vec<u32> { ... }
}

Trait for assembling functionalities.

Required methods

fn assemble_into(&self, result: &mut Vec<u32>)

Assembles the current object into the result vector, reducing the need for lots of allocations

Provided methods

fn assemble(&self) -> Vec<u32>

Assembles the current object and returns the binary code. Helper method to remain backwards compatible, calls assemble_into