rspirv::binary

Trait Assemble

pub trait Assemble {
    // Required method
    fn assemble_into(&self, result: &mut Vec<u32>);

    // Provided method
    fn assemble(&self) -> Vec<u32> { ... }
}

Trait for assembling functionalities.

Required Methods

fn assemble_into(&self, result: &mut Vec<u32>)

Assembles the current object into the result vector, reducing the need for lots of allocations

Provided Methods

fn assemble(&self) -> Vec<u32>

Assembles the current object and returns the binary code. Helper method to remain backwards compatible, calls assemble_into