rspirv::grammar
Module reflect

Reflect functions for SPIR-V instructions.

Functions

is_abort

Returns true if the given opcode aborts execution.

is_annotation

Returns true if the given opcode is for an annotation instruction.

is_block_terminator

Returns true if the given opcode is for a terminator instruction.

is_branch

Returns true if the given opcode is a branch instruction.

is_constant

Returns true if the given opcode is for a constant-defining instruction.

is_debug

Returns true if the given opcode is for a debug instruction.

is_location_debug

Returns true if the given opcode is for a location debug instruction.

is_nonlocation_debug

Returns true if the given opcode is for a non-location debug instruction.

is_return

Returns true if the given opcode is a return instruction.

is_return_or_abort

Returns true if the given opcode is a return instruction or it aborts execution.

is_type

Returns true if the given opcode is for a type-declaring instruction.

is_variable

Returns true if the given opcode is for a variable-defining instruction.